First real commit of v20
There has been so much changed to bring it to v20 but I am tired of having all these changes not pushed. I know the consequences and I will not be responsible this.

MOSTLY functional. Some things are still broken and incomplete. This is mostly for test build purposes. I have a v19 backup elsewhere, I can calm down.